WEB-3770 move getReportDaysText to different file
